Use Decimal for Numeric columns and fix ERRLOG_IX1 DESC ordering
- Change Mapped[float] to Mapped[Decimal] on all financial Numeric
  columns (portfolio.py, transaction_history.py, poshist.py) to
  preserve exact decimal semantics from DB2 DECIMAL types.
- Add DESC ordering on error_severity in ERRLOG_IX1 index to match
  the original DB2 DDL (ERRLOG.sql line 43).
- Update Alembic migration and tests accordingly.

+++ python/src/database/db2/portfolio.py
@@ -4,6 +4,7 @@ Migrated from: src/database/db2/db2-definitions.sql
 """
 
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import Date, DateTime, ForeignKey, Index, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -57,13 +58,13 @@ class InvestmentPositions(Base):
     )
     investment_id: Mapped[str] = mapped_column(String(10), primary_key=True)
     position_date: Mapped[date] = mapped_column(Date, primary_key=True)
-    quantity: Mapped[float] = mapped_column(
+    quantity: Mapped[Decimal] = mapped_column(
         Numeric(precision=18, scale=4), nullable=False
     )
-    cost_basis: Mapped[float] = mapped_column(
+    cost_basis: Mapped[Decimal] = mapped_column(
         Numeric(precision=18, scale=2), nullable=False
     )
-    market_value: Mapped[float] = mapped_column(
+    market_value: Mapped[Decimal] = mapped_column(
         Numeric(precision=18, scale=2), nullable=False
     )
     currency_code: Mapped[str] = mapped_column(String(3), nullable=False)
